+++ platformio/commands/list.py
@@ -9,7 +9,7 @@ from platformio.pkgmanager import PackageManager
 @command("list", short_help="List installed platforms")
 def cli():
 
-    for name, pkgs in PackageManager.get_installed().iteritems():
+    for name, pkgs in PackageManager.get_installed().items():
         echo("{name:<20} with packages: {pkgs}".format(
             name=style(name, fg="cyan"),
             pkgs=", ".join(pkgs.keys())