+++ platformio/commands/uninstall.py
@@ -5,10 +5,10 @@ from click import argument, command, secho
 
 from platformio.exception import PlatformNotInstalledYet
 from platformio.pkgmanager import PackageManager
-from platformio.platforms._base import PlatformFactory
+from platformio.platforms.base import PlatformFactory
 
 
-@command("uninstall", short_help="Uninstall the platforms")
+@command("uninstall", short_help="Uninstall platforms")
 @argument("platforms", nargs=-1)
 def cli(platforms):