+++ platformio/platforms/atmelavr.py
@@ -3,7 +3,7 @@
 
 from os.path import join
 
-from platformio.platforms._base import BasePlatform
+from platformio.platforms.base import BasePlatform
 
 
 class AtmelavrPlatform(BasePlatform):
@@ -16,12 +16,14 @@ class AtmelavrPlatform(BasePlatform):
 
         "toolchain-atmelavr": {
             "path": join("tools", "toolchain"),
+            "alias": "toolchain",
             "default": True
         },
 
         "tool-avrdude": {
             "path": join("tools", "avrdude"),
-            "default": True,
+            "alias": "uploader",
+            "default": True
         },
 
         "framework-arduinoavr": {