+++ HISTORY.rst
@@ -1,6 +1,26 @@
 Release History
 ===============
 
+1.0.0 (?)
+---------
+
+0.4.0 (2014-07-31)
+------------------
+
+* Implemented ``serialports`` command
+* Allowed to put special build flags only for ``src`` files via
+  ``srcbuild_flags`` environment option
+* Allowed to override some of settings via system environment variables
+  such as: ``$PIOSRCBUILD_FLAGS`` and ``$PIOENVS_DIR``
+* Added ``--upload-port`` option for ``platformio run`` command
+* Implemented (especially for `SmartAnthill <http://smartanthill.ikravets.com/>`_)
+  ``platformio run -t uploadlazy`` target (no dependencies to framework libs,
+  ELF and etc.)
+* Allowed to skip default packages via ``platformio install --skip-default-package`` flag
+* Added tools for Raspberry Pi platform
+* Added support for Microduino and Raspduino boards in ``atmelavr`` platform
+
+
 0.3.1 (2014-06-21)
 ------------------